+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Automatic retries for queued ingestion
 * `executeQuery`, `executeMgmt` to call with a specific type.
+### Fixed 
+* Timer was used if authentication throws after client was closed 
 
 ## [5.0.0]
 ### Fixed