+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [5.0.1]
 ### Added
 * Automatic retries for queued ingestion
 * `executeQuery`, `executeMgmt` to call with a specific type.
 ### Fixed 
 * Timer was used if authentication throws after client was closed 
 * Public client credentials (user prompt,device auth) are synchronized - so that users are prompt once
+* Msal version was old after some change and collide with azure-identity msal dependency
 
 ## [5.0.0]
 ### Fixed